# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation's Employee Manager",
        "title": "Employee Manager",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Bui Viet Hoang",
            "email": "mynamebvh@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/departments": {
            "post": {
                "description": "Create new department on the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Create new department",
                "parameters": [
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/department.NewDepartment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Department"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/department.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/department.MessageResponse"
                        }
                    }
                }
            }
        },
        "/departments/{department_id}": {
            "get": {
                "description": "Get department by ID on the system",
                "tags": [
                    "department"
                ],
                "summary": "Get department by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of department",
                        "name": "department_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Department"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/department.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/department.MessageResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update department by ID on the system",
                "tags": [
                    "department"
                ],
                "summary": "Update department by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of department",
                        "name": "department_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/department.NewDepartment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Department"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/department.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/department.MessageResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete department by ID on the system",
                "tags": [
                    "department"
                ],
                "summary": "Delete department by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of department",
                        "name": "department_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/department.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/department.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/department.MessageResponse"
                        }
                    }
                }
            }
        },
        "/requests": {
            "get": {
                "description": "Get list request",
                "tags": [
                    "request"
                ],
                "summary": "Get list request",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.PaginationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.MessageResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "request"
                ],
                "summary": "Create new request",
                "parameters": [
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.NewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Request"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.MessageResponse"
                        }
                    }
                }
            }
        },
        "/requests/{request_id}": {
            "get": {
                "description": "Get request by id",
                "tags": [
                    "request"
                ],
                "summary": "Get request by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of request",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.MessageResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update request by ID on the system",
                "tags": [
                    "request"
                ],
                "summary": "Update request by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of request",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.NewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Request"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "department.MessageResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "department.NewDepartment": {
            "type": "object",
            "required": [
                "address",
                "department_code",
                "name",
                "status"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Tầng 3"
                },
                "department_code": {
                    "type": "string",
                    "example": "D5"
                },
                "name": {
                    "type": "string",
                    "example": "Phòng D5"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.Department": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "department_code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "models.Request": {
            "type": "object",
            "properties": {
                "approved_by": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "request.MessageResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "request.NewRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "request.PaginationResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "page_current": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "page_total": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Authentication": {
            "description": "Bearer token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}